fix: better handle broken punctuation caused by `<br>` elements
Annotated lyrics from genius are formatted as anchor elements, where the
text is inside span elements, with `<br>` elements for newlines. Those
`<br>` elements don't persist when parsing html into a `\DOMDocument`.
This was causing the separation between sentences to be lost in the
final result. This adds a regex search-and-replace, where it adds a
literal everywhere where an upper-case letter follows directly after a
lower-case letter.